Rename simulators to sim
