+++ building3d/sim/rays/config.py
