+++ building3d/sim/rays/dump_buffers.py
