+++ building3d/sim/rays/find_nearby_polygons.py
