+++ building3d/sim/rays/find_transparent.py
