+++ building3d/sim/rays/movie_from_buffer.py
