+++ building3d/sim/rays/ray_buff_plotter.py
