+++ building3d/sim/rays/simulation.py
