+++ building3d/sim/rays/voxel_grid.py
